Cleaned up the activity scoring code.

+++ techtribes-core/src/je/techtribes/util/comparator/ActivityByScoreComparator.java
@@ -12,7 +12,7 @@ public class ActivityByScoreComparator implements Comparator<Activity> {
 
     @Override
     public int compare(Activity activity1, Activity activity2) {
-        int result = new Double(activity2.getScore()).compareTo(activity1.getScore());
+        int result = new Long(activity2.getScore()).compareTo(activity1.getScore());
         if (result != 0) {
             return result;
         } else {
